Decouple reactContext from AmazonListener
