+++ IapExample/android/build.gradle
@@ -34,6 +34,17 @@ buildscript {
 
 allprojects {
     repositories {
+        // Fix to react-native not being found: https://github.com/facebook/react-native/issues/35204#issuecomment-1304740228
+        exclusiveContent {
+           filter {
+               includeGroup "com.facebook.react"
+           }
+           forRepository {
+               maven {
+                   url "$rootDir/../node_modules/react-native/android"
+               }
+           }
+       }
         maven {
             // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
             url("$rootDir/../node_modules/react-native/android")
