+++ android/src/amazon/java/com/dooboolab/RNIap/EventSender.kt
@@ -1,0 +1,10 @@
+package com.dooboolab.RNIap
+
+import com.facebook.react.bridge.WritableMap
+
+interface EventSender {
+    fun sendEvent(
+        eventName: String,
+        params: WritableMap?
+    )
+}
